build: float not double for CenteredLayout::make_fmt line_height_factor
Page::Format::line_height_factor is float; passing a double through
make_fmt and then assigning to the float field is a narrowing
conversion. paper_s3 builds with -Werror=narrowing, so this fails
to compile. (My local sandbox's kconfgen is broken so I can't pio-
build; user caught it post-flash.)

Three sites:
- include/viewers/centered_layout.hpp: parameter declaration
- src/viewers/centered_layout.cpp: parameter definition
- src/viewers/sleep_screen_viewer.cpp: LINE_HEIGHT_FACTOR constant
- src/viewers/usb_msc_viewer.cpp: LINE_HEIGHT_FACTOR constant

All four updated to float (with `f` suffix on the literals to
avoid the same trap on the constants themselves). Float is also
sufficient precision for the actual values used (1.0, 1.1, 1.15) —
they only need one decimal digit.

Header doc comment now flags the narrowing constraint so a future
caller passing a double literal gets pointed at the right shape
before they hit the build.

+++ include/viewers/centered_layout.hpp
@@ -30,11 +30,18 @@ namespace CenteredLayout {
   // Build a Page::Format suitable for centered full-screen text.
   // line_height_factor is the only knob that varied between
   // sleep_screen (1.1) and usb_msc (1.15); pass yours.
+  //
+  // line_height_factor is `float` (not double) to match Page::
+  // Format::line_height_factor exactly — the paper_s3 build runs
+  // with -Werror=narrowing, so passing a double through here and
+  // assigning it to a float field fails to compile. Float is also
+  // sufficient for the actual values used (one decimal of
+  // precision).
   Page::Format make_fmt(int8_t           font_index,
                         int16_t          font_size,
                         Fonts::FaceStyle style,
                         CSS::Align       align,
-                        double           line_height_factor);
+                        float            line_height_factor);
 
   // Paint `text` centered horizontally at panel-y `y`. Mutates
   // fmt.align to CENTER (the prior behavior of both private
